reduce layer to 5 from 10, use simple hand algorithm

+++ tasks/gen_format.py
@@ -2,8 +2,8 @@ import os
 import shutil
 
 pwd = os.path.dirname(os.path.realpath(__file__))
-hierarchy_path = os.path.join(pwd, "..", "..", "new_templates")
-target_path = os.path.join(pwd, "..", "..", "large_template_front")
+hierarchy_path = os.path.join(pwd, "..", "..", "playground_cat_template")
+target_path = os.path.join(pwd, "..", "..", "playground_templates")
 if not os.path.exists(target_path):
     os.makedirs(target_path)
 
@@ -17,7 +17,7 @@ if __name__ == '__main__':
             label = str(abs(int(subdir)))
             for filename in os.listdir(full_subdir):
                 pwf = os.path.join(full_subdir, filename)
-                if os.path.isfile(pwf):
+                if os.path.isfile(pwf) and filename != ".gitignore":
                     new_name = str(counter) + "." + str(label) + "." + str(0 if int(subdir) > 0 else -1) + ",0,100,100.png"
                     shutil.copy(pwf, os.path.join(target_path, new_name))
                     counter = counter + 1
